Django/1013 Django Auth 2/requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
black==22.8.0
click==8.1.3
colorama==0.4.5
Django==3.2.13
django-appconf==1.0.5
django-bootstrap-v5==1.0.11
django-bootstrap5==22.1
django-extensions==3.2.1
django-imagekit==4.1.0
mypy-extensions==0.4.3
pathspec==0.10.1
pilkit==2.0
Pillow==9.2.0
platformdirs==2.5.2
pytz==2022.2.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tomli==2.0.1
